% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plot.R
\name{convert_ts_forecast_to_df}
\alias{convert_ts_forecast_to_df}
\title{convert_ts_forecast_to_df}
\usage{
convert_ts_forecast_to_df(forecast)
}
\arguments{
\item{forecast}{timeseries object to convert}
}
\description{
method for converting a timerseries to a dataframe so that it can be plotted with ggplot2 and keep a Date x-axis.
}
