% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flickr_userdays}
\alias{flickr_userdays}
\title{Popularity of Flickr, in User-Days}
\format{
A time series object with 156 observations.
}
\source{
Flickr. (2019). Retrieved October, 2019, from https://flickr.com/
}
\usage{
flickr_userdays
}
\description{
A time series representing the popularity of Flickr in the United States, as measured in user-days. Here, user-days count the number of unique users posting on Flickr on a given day.
}
\keyword{datasets}
