% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_predict.R
\name{generate_proxy_trend_forecasts}
\alias{generate_proxy_trend_forecasts}
\title{Generate Proxy Trend Forecasts}
\usage{
generate_proxy_trend_forecasts(
  object,
  n_ahead,
  starttime,
  endtime,
  proxy_trend_correction,
  ts_frequency
)
}
\arguments{
\item{object}{A visitation model object.}

\item{n_ahead}{The number of desired forecasts.}

\item{starttime}{The start time of the desired forecasts.}

\item{endtime}{The end time of the desired forecasts.}

\item{proxy_trend_correction}{The lag correction needed on the proxy trend.}

\item{ts_frequency}{Frequency of the time series to forecast.}
}
\value{
A time series object storing forecasts for the proxy trend.
}
\description{
Generating proxy trend forecasts from objects of the class "visitation_model".
}
