% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_visitation_model.R
\name{imputation}
\alias{imputation}
\title{Imputation}
\usage{
imputation(x)
}
\arguments{
\item{x}{A numeric vector (usually the log visitation counts or photo-user days).}
}
\value{
A numeric vector with the negative infinities replaced with appropriate numbers.
}
\description{
Imputation by replacing negative infinities with appropriate numbers.
}
