% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{label_visitation_forecast}
\alias{label_visitation_forecast}
\title{labeled_visitation_forecast Class}
\usage{
label_visitation_forecast(visitation_forecast, label)
}
\arguments{
\item{visitation_forecast}{A visitation_forecast object}

\item{label}{A character string of the label of forecast}
}
\value{
Object of class "visitation_forecast_ensemble".
}
\description{
Class for visitation_model predictions (for use with predict.visitation_model()).
}
