% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_decomposition}
\alias{new_decomposition}
\title{"decomposition" Constructor Function}
\usage{
new_decomposition(reconstruction_list, grouping_matrix, window_length, ts_ssa)
}
\arguments{
\item{reconstruction_list}{A list containing important information about the reconstructed time series. In particular, it contains the reconstructed main trend component, overall trend component, seasonal component for each period specified in suspected_periods, and overall seasonal component.}

\item{grouping_matrix}{A matrix containing information about the locations of the eigenvalue groups for each period in suspected_periods and trend component. The locations are indicated by '1'.}

\item{window_length}{A numeric indicating the window length.}

\item{ts_ssa}{An object of the class "ssa".}
}
\value{
A list of the class "decomposition".
}
\description{
Constructs objects of the "decomposition" class.
}
