% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_visitation_forecast_ensemble}
\alias{new_visitation_forecast_ensemble}
\title{visitation_forecast_ensemble Class}
\usage{
new_visitation_forecast_ensemble(visitation_forecasts, labels)
}
\arguments{
\item{visitation_forecasts}{An array of  visitation_forecast object}

\item{labels}{An array of  labels associated with visitation_forecast}
}
\description{
Class for plotting an array of visitation_forecast objects
}
