% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_predict.R
\name{prediction_warning}
\alias{prediction_warning}
\title{Notify User prediction warning on constant is 0}
\usage{
prediction_warning(constant)
}
\arguments{
\item{constant}{The B_0 parameter of the model.}
}
\value{
No return value
}
\description{
Notfy the user of details related to the outputs of the model being potentially inaccurate when constant of model is 0.
}
