% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_print.R
\name{print.visitation_model}
\alias{print.visitation_model}
\title{visitation_model Summary Method}
\usage{
\method{print}{visitation_model}(x, ...)
}
\arguments{
\item{x}{An object of class "decomposition".}

\item{...}{Additional arguments.}
}
\value{
A "decomposition" class object.
}
\description{
Methods for summarizing objects of the class "decomposition".
}
\examples{
#Example:

data("park_visitation")
data("flickr_userdays")

n_ahead <- 12
park <- "YELL"
pud_ts <- ts(park_visitation[park_visitation$park == park,]$pud, start = 2005, freq = 12)
pud_ts <- log(pud_ts)
trend_proxy <- log(flickr_userdays)

vm <- visitation_model(pud_ts,trend_proxy)
summary(vm)
}
