% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_visitation_model.R
\name{trim_training_data}
\alias{trim_training_data}
\title{trim training data}
\usage{
trim_training_data(onsite_usage = NULL, ref_series = NULL)
}
\arguments{
\item{onsite_usage}{A vector which stores monthly on-site usage for a particular social media platform and recreational site.}

\item{ref_series}{A numeric vector specifying the original visitation series. The default option is NULL, implying that no such series is available. If such series is available, then its length must be the same as that of \code{onsite_usage}.}
}
\value{
a list of onsite_usage and ref_series that has been trimmed and modified to share same window of time.
}
\description{
Makes sure that the provided onsite_usage and ref_series have at least 12 counts and overlap.
}
