% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_visitation_model.R
\name{yearsToMonths}
\alias{yearsToMonths}
\title{Converting Annual Counts into Monthly Counts}
\usage{
yearsToMonths(visitation_years, pud)
}
\arguments{
\item{visitation_years}{A numeric vector with annual visitation counts. If not available, NA should be entered.}

\item{pud}{A numeric vector for the monthly photo-user-days corresponding to \code{visitation_years}. As such, the length of \code{pud} needs to be exactly 12 times as long as \code{visitation_counts}.}
}
\value{
A numeric vector with estimated monthly visitation counts based on the annual counts and monthly photo-user-days.
}
\description{
Convert annual counts into monthly counts using photo-user-days.
}
